refactor(front): moved handleDecsionFinish from LateralBar to GameScreen to be able to clean the state
